move hover threshold logic up to RepoContainer
