+++ client/web/src/components/shared.tsx
@@ -9,18 +9,9 @@ import {
 import { isErrorLike } from '../../../shared/src/util/errors'
 import { HoverOverlay, HoverOverlayProps } from '../../../shared/src/hover/HoverOverlay'
 import { useLocalStorage } from '../util/useLocalStorage'
+import { HoverThresholdProps } from '../repo/RepoContainer'
 
 // Components from shared with web-styling class names applied
-
-export const HOVER_COUNT_KEY = 'hover-count'
-export const HOVER_THRESHOLD = 5
-
-export interface HoverThresholdProps {
-    /**
-     * Called when the threshold of hovers with a content is reached.
-     */
-    onHoverThresholdReached?: () => void
-}
 
 export const WebHoverOverlay: React.FunctionComponent<HoverOverlayProps & HoverThresholdProps> = props => {
     const [dismissedAlerts, setDismissedAlerts] = useLocalStorage<string[]>('WebHoverOverlay.dismissedAlerts', [])
@@ -42,26 +33,16 @@ export const WebHoverOverlay: React.FunctionComponent<HoverOverlayProps & HoverT
     }
 
     const { hoverOrError } = propsToUse
-    const { onHoverThresholdReached, hoveredToken } = props
+    const { onHoverShown, hoveredToken } = props
 
     /** Whether the hover has actual content (that provides value to the user) */
     const hoverHasValue = hoverOrError !== 'loading' && !isErrorLike(hoverOrError) && !!hoverOrError?.contents?.length
 
     useEffect(() => {
         if (hoverHasValue) {
-            const count = parseInt(localStorage.getItem(HOVER_COUNT_KEY) ?? '0', 10) + 1
-
-            if (count > HOVER_THRESHOLD) {
-                return
-            }
-
-            if (count === HOVER_THRESHOLD) {
-                onHoverThresholdReached?.()
-            }
-
-            localStorage.setItem(HOVER_COUNT_KEY, count.toString(10))
+            onHoverShown?.()
         }
-    }, [hoveredToken?.filePath, hoveredToken?.line, hoveredToken?.character, onHoverThresholdReached, hoverHasValue])
+    }, [hoveredToken?.filePath, hoveredToken?.line, hoveredToken?.character, onHoverShown, hoverHasValue])
 
     return (
         <HoverOverlay
